+++ src/qnethernet_opts.h
@@ -64,6 +64,12 @@
 #define QNETHERNET_ENABLE_PROMISCUOUS_MODE 0
 #endif
 
+// Enables raw frame loopback when the destination MAC address matches the local
+// MAC address.
+#ifndef QNETHERNET_ENABLE_RAW_FRAME_LOOPBACK
+#define QNETHERNET_ENABLE_RAW_FRAME_LOOPBACK 0
+#endif
+
 // Enables raw frame support.
 #ifndef QNETHERNET_ENABLE_RAW_FRAME_SUPPORT
 #define QNETHERNET_ENABLE_RAW_FRAME_SUPPORT 0